backend/requirements.txt
Django==4.2.24
djangorestframework==3.16.1
python-dotenv==1.0.0
httpx==0.27.2
psycopg2-binary==2.9.9
